System.Console.SimpleLineEditor

Description

A simple line editor, using the GNU readline library if available, or a small emulation otherwise.

Documentation

initialise :: IO ()

Set up the environment so that the terminal passes characters directly into the Haskell program, for immediate interpretation by the line editor.

restore :: IO ()

Restore the environment so that the terminal is usable in normal mode once again.

getLineEdited :: String -> IO (Maybe String)

delChars :: String -> IO ()

Remove the given string from immediately behind (to the left of) the current cursor position.